requirements.txt
os
random
discord
datetime
asyncio
Pillow
python-dotenv
